Modifications needed to switch the Albedo texture to a TextureArray. Note that the other textures (_Metallic, _BumpMap,...) are still regular Texture2D, not Texture2DArrays.
Shader has been renamed to MeshBaker/StandardTextureArray so that it won't conflict with the regular standard shader.
#include statements have been modified so that .cginc import files are replace with my modified versions for "StandardCore", "StandardCoreForward", "StandardInput"
_MainTex type has been switched from "2D" to "2DArray"
_MainTex is declared using: UNITY_DECLARE_TEX2DARRAY(_MainTex)
It is expected that the mesh UV0 coordinates will be a float3 instead of float 2. The third coordinate is the slice index that should be used in the Texture Array.
uv0 coordinate is now float3 instead of float2. The 3rd coordinate is the slice index that should be used in the Texture Array. This coordinate is assigned from the vertex UVs.
Calls to SpecularGloss, MetallicGloss, and Alpha now take float3 instead of float2